requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
pycodestyle==2.10.0
pydantic==1.10.5
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wolframclient==1.1.7
yarl==1.8.2
